List and Map Scan Implementation
Adds a shared RepeatedColumnReader superclass that captures the mapping of container level rows to nested rows.

Adds PositionalFilter to represent situations where different positions in a column have different filters. This occurs when filters apply to values at a specific subscript in a list or map.

Adds an ordinal position to StructFilter when it occurs inside another StructFilter. If different StructFilters apply to different positions in a struct column inside a list or map, we need to distinguish between the different StructFilter children of the StructFilter of the container list/map.

Adds an exception for missing subscript. This is added to an output qualifying set of a list or map reader for rows where all the subscripts for which a filter was provided did not exist and all the filters whose subscripts did exist passed.