Add common scalar reads to OrcInputStream
Adds an optimized read path for varints, floats and doubles. This
avoids copying and bounds checking for consecutive bytes.

Removes the use of Slice from OrcInputStream. Keeps the decompression
buffer separate from the Slice read from the compressed input. The
backing byte[] of the compressed input Slice is directly used as the
buffer when there is no compression, avoiding needless copy.