Make updateTaskStatus not synchronized within SqlStageExecution
This synchronization is a contention point and introduces a
dead lock for recoverable execution.

The `updateTaskStatus` method only alters the `failedTasks` and
`finishedTasks` variables.

The `failedTasks` list is only used to check whether there are too
many tasks failed, and whether the next recovery is possible. It
is always updated and checked within the same method. Thus,
synchronization is not needed, as long as the `failedTasks` set is
concurrent (allows concurrent modifications).

The `finishedTasks` is used to check whether the stage execution is
done. It is updated only withing the `updateTaskStatus`, and it is
checked in 2 different places. The `updateTaskStatus` and the
`schedulingComplete`. As long as the `finishedTasks` allows concurrent
modifications itself the overall lock is not needed. As:

- In the `schedulingComplete`:
  `transitionToScheduled` always happens before the
  `finishedTasks.size() == allTasks.size()`
- In the `updateTaskStatus`:
  `finishedTasks.add(taskStatus.getTaskId())` always happens before
  the `stageExecutionState == StageExecutionState.SCHEDULED`
  and the `finishedTasks.size() == allTasks.size()`