Fix getTableStatistics with subfield pruning
ColumnHandles contain output columns, while predicateColumns contain columns
used in filter. These two collections overlap when an output column is used in
a filter. ColumnHandles may also include subfield pruning information
(requiredSubfields), but predicateColumns never do. When output column
contains subfield pruning information and is used in a filter, columnHandles
and predicateColumns include the same column, but corresponding ColumnHandle
instances don't match. This causes allColumnHandles to include the column
twice and leads to java.lang.IllegalArgumentException: Multiple entries with
same key.

The fix is to add only a subset of predicateColumns that are not present in
columnHandles to allColumnHandles.

Also, make sure that filters that use non-trivial expressions or udfs still
get an estimate based on UNKNOWN_FILTER_COEFFICIENT.