Evaluate constant column dependent filter functions only once
Track if a filter function depends on constant columns only. If so,
evaluate it on a single row. If this is false then the whole split
will come out empty. If on the other hand, all filter functions pass
for at least one row, we know that all constant dependent filter
functions also passed and will not need to evaluate them again within
this split.